Add files via upload
A Python program that helps you decide where to eat by randomly selecting a restaurant option using if statements. The program eliminates the hassle of choosing where to go or what to order by offering a fun and simple way to pick a random restaurant from a list of options.